Change RequestId and ProgressToken to use int64_t instead of int (#121)
JSON-RPC allows string or number for request IDs. Using int64_t instead
of int prevents overflow when generating numeric request IDs from
uint64_t counters.

Changes:
- Update RequestId type from variant<std::string, int> to variant<std::string, int64_t>
- Update ProgressToken type to use int64_t for consistency
- Add factory function for int64_t request IDs
- Update all usages of holds_alternative<int> to holds_alternative<int64_t>
- Update all usages of get<int> to get<int64_t> for RequestId/ProgressToken
- Change pending_requests_ map key from int to int64_t in RequestTracker
- Cast uint64_t request ID counters to int64_t to avoid overflow

+++ include/mcp/server/mcp_server.h
@@ -694,7 +694,7 @@ class McpServer : public application::ApplicationBase,
     // Convert RequestId to string key
     std::string key = holds_alternative<std::string>(id)
                           ? get<std::string>(id)
-                          : std::to_string(get<int>(id));
+                          : std::to_string(get<int64_t>(id));
     auto it = pending_requests_.find(key);
     return (it != pending_requests_.end() && it->second->cancelled.load());
   }
